move list context to new file
