+++ src/components/chat.tsx
@@ -2,7 +2,8 @@ import React, { FC, useState, useEffect } from 'react'
 import type { ChatCompletionRequestMessage as Message } from 'openai'
 import ModelDropdown from '@/components/model-dropdown'
 import MessageInput from '@/components/message-input'
-import MessageList, { ListContext } from '@/components/message-list'
+import MessageList from '@/components/message-list'
+import ListContext from '@/lib/list-context'
 import type { TreeNode } from '@/lib/message-tree'
 import * as tr from '@/lib/message-tree'
 import { jsonPostBody } from '@/lib/fetch'
